seperated scrape.py for more spiders

+++ R/data/scrapers/cache.py
@@ -16,7 +16,8 @@ def write(response, dirname):
 
     filepath   = os.path.join(dirdata, filename)
 
-    with open(filepath, 'wb') as f:
-        buffrr = response.body
+    if not os.path.isfile(filepath):
+        with open(filepath, 'wb') as f:
+            buffrr = response.body
 
-        f.write(buffrr)
+            f.write(buffrr)
